% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_tabcon.R
\name{app.tabcon}
\alias{app.tabcon}
\title{La fonction "app.tabcon"}
\usage{
app.tabcon(
  dataframe = databool.N1[, names(databool.N1) \%w/o\% paste("clust.c1.", 2:5, sep =
    "")],
  var.base = "clust.c1.1",
  var.exclude = NULL,
  decim = 1
)
}
\arguments{
\item{dataframe}{le data.frame}

\item{var.base}{nom (char) ou index de la variable à tester}

\item{var.exclude}{NULL optional: nom ou index des variables à exclure}

\item{decim}{1 nombre de décimales dans le résultat (round())}

\item{type}{1 what is it?}
}
\value{
list("Effectifs"=tab1.1, "Frequence"=tab2.1, "Freq_ligne"=tab3.1, "Freq_col"=tab4.1)
}
\description{
cré des tables de contingences (effectifs et fréquences*3) sur un data.frame entier pour une variable de base
}
\references{
table() ; prop.table() ; round()
}
\author{
Elie
}
