% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_test.R
\name{app.test}
\alias{app.test}
\title{La fonction "app.test"}
\usage{
app.test(
  dataframe = databool.N1[, names(databool.N1) \%w/o\% paste("clust.c1.", 2:5, sep =
    "")],
  var.base = "clust.c1.1",
  var.test = NULL,
  var.exclude = NULL,
  test.t = "fisher"
)
}
\arguments{
\item{dataframe}{le data.frame}

\item{var.base}{nom (char) ou index de la variable à tester}

\item{var.test}{NULL optionnal: nom ou index des variables à tester contre var.base}

\item{var.exclude}{NULL optional: nom ou index des variables à exclure du test}
}
\value{
list("res.test"=res.test, "res.df"=res.df)
}
\description{
Applique un test d'indépendance (Khi2 ou fisher) entre une variable d'un data.frame et toutes les autres (ou certaines) variables
}
\examples{
app.test<-function(dataframe=databool.N1[, names(databool.N1)\%w/o\%paste("clust.c1.", 2:5, sep="")],
var.base="clust.c1.1",
var.test=NULL,
var.exclude=NULL,
test.t="fisher"
)
}
\references{
voir fisher.test() et chisq.test()
}
\author{
Elie
}
