% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asindice_df.R
\name{asindice.df}
\alias{asindice.df}
\title{asindice.df applique asindice sur une data.frame ou une matrice}
\usage{
asindice.df(
  DF = NAF17ens,
  Margin = 2,
  type = "VALEUR",
  rowbase = 1,
  indice = 100,
  alter.rownames = NULL,
  alter.colnames = NULL,
  change.nom = NULL
)
}
\arguments{
\item{DF}{une data.frame ou matrice}

\item{Margin}{le temps en ligne ou en colonne}

\item{type}{'VALEUR' ou 'TAUX': renseigne le type de variable entrante}

\item{rowbase}{le rang de la base dans la série temporelle}

\item{indice}{la base (numérique)}

\item{alter.rownames}{characters vector}

\item{change.nom}{applique ou non le chaine de caracteres change.nom aux noms de variables pré-existants}
}
\value{
un vecteur res01
}
\description{
asindice.df applique asindice sur une data.frame ou une matrice
}
