% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asindice.R
\name{asindice}
\alias{asindice}
\title{asindice crée un indice à paretir d'une variable numérique. Précise la base par le numéro de ligne et indice à choisir (100, 1000...)}
\usage{
asindice(type = "VALEUR", timeserie = emploiEQTP$TOTAL, rowbase = 1,
  indice = 100)
}
\arguments{
\item{type}{'VALEUR' ou 'TAUX': renseigne le type de variable entrante}

\item{timeserie}{la série temporelle}

\item{rowbase}{le rang de la base dans la série temporelle}

\item{indice}{la base (numérique)}
}
\value{
un vecteur res01
}
\description{
asindice crée un indice à paretir d'une variable numérique. Précise la base par le numéro de ligne et indice à choisir (100, 1000...)
}
