% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bool_df.R
\name{create.bool.df}
\alias{create.bool.df}
\title{La fonction "create.bool.df"}
\usage{
create.bool.df(dataframe = dataclust)
}
\arguments{
\item{dataframe}{un data.frame}
}
\value{
un data.frame
}
\description{
Cré un data.frame avec autant de variables booléennes que de modalités présentes dans toutes les variables de type factor dans le data.frame d'origine
}
\examples{
create.bool.df(dataframe= dataclust)
}
\references{
no
}
\author{
Elie
}
