% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_sup_5.R
\name{eff_sup_5}
\alias{eff_sup_5}
\title{La fonction "eff_sup_5"}
\usage{
eff_sup_5(var1, var2)
}
\arguments{
\item{var1}{la premiere variable : typa factor}

\item{var2}{la deuxième variable : type factor}
}
\value{
booleen: true si effectif supérieur à 5 partout
}
\description{
Verifie que l'effectif est bien supérieur à 5 dans chaque case du tableau de contingence entre deux variables
Permet de savoir si peut appliquer un test du Chi2.
}
\examples{
eff_sup_5<-function(var1=as.factor(infert$spontaneous),
var2=as.factor(infert$education))
}
\references{
voir table()
}
\author{
Romane
}
