% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_cu.R
\name{freq.cu}
\alias{freq.cu}
\title{La fonction "freq.cu"}
\usage{
freq.cu(df = propomat, var = "X2", debut = "haut")
}
\arguments{
\item{df}{le dataframe. Si df est une table ou une matrice, sera transformé en data.frame}

\item{var}{la variable d'intéret dans le data.frame en chaine de caracteres ou en index}

\item{debut}{de bas en haut ('bas') ou de haut en bas('haut')}
}
\value{
un vecteur res
}
\description{
freq.cu calcul les fréquences cumulées pour une variable d'un df, de bas en haut ('bas') ou de haut en bas('haut')
}
