% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multsheet.R
\name{multsheet}
\alias{multsheet}
\title{multsheet permet de lire plusieurs feuille d'une meme table xls avec la fonction read_xls}
\usage{
multsheet(
  tablepath = "secteur_dep.xls",
  sheet.vec = 1:20,
  sheet.name = NULL,
  SKIP = 4,
  aslist = 1
)
}
\arguments{
\item{tablepath}{"chemin/du/fichiers.xls"}

\item{sheet.vec}{vecteur numérique des feuilles à impotrer (exemple: 1:20)}

\item{sheet.name}{nom affecté à chaque data.frame créée (suivi de son numéro)}

\item{SKIP}{un seul numérique ou un vecteur de numérique de meme longueur que sheet.vec}

\item{aslist}{1=return a list of data.frame ; 2= enregistre les dataframe dans l'env ; 3= combo!}
}
\value{
rien
}
\description{
multsheet permet de lire plusieurs feuille d'une meme table xls avec la fonction read_xls
}
