% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMES.R
\name{NAMES}
\alias{NAMES}
\title{NAMES: data.frame de names d'un data.frame}
\usage{
NAMES(dataframe = data.o)
}
\arguments{
\item{dataframe}{le data.frame}
}
\value{
res, un data.frame
}
\description{
NAMES: data.frame de names d'un data.frame
}
