% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom_wo.R
\name{phantom.wo}
\alias{phantom.wo}
\title{returns a logical vector indicating if there is not a match for its left operand}
\usage{
phantom.wo(x = "P:/Documents_de_Romane/wo.R")
}
\value{
nothing
}
\description{
returns a logical vector indicating if there is not a match for its left operand
}
\note{
phantom
}
\author{
Elie
}
