% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resum_factors.R
\name{resum.factors}
\alias{resum.factors}
\title{resum.factors retourne un résumé d'une variable qualitative (effectifs et fréquences) dans deux tableaux}
\usage{
resum.factors(dim = data_log$contrepartie, decim = 0)
}
\arguments{
\item{dim}{variable factor (quali)}

\item{decim}{nombre de décimales pour l'output}
}
\value{
une matrice res
}
\description{
resum.factors retourne un résumé d'une variable qualitative (effectifs et fréquences) dans deux tableaux
}
