% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_con.R
\name{table.con}
\alias{table.con}
\title{table.con retourne des tables de contingences pour deux variables d'un même data.frame. Effectifs, fréquence totale, en ligne et en colonne}
\usage{
table.con(
  data = df,
  var = c(),
  decim = 0,
  na.pattern = c("", " ", NA),
  na.recode = "non-réponse",
  ...
)
}
\arguments{
\item{data}{le data.frame}

\item{var}{vecteur des 2 variables d'intéret. Chaines de caracteres.}

\item{decim}{nombre de décimales}
}
\value{
res, un data.frame
}
\description{
table.con retourne des tables de contingences pour deux variables d'un même data.frame. Effectifs, fréquence totale, en ligne et en colonne
}
