% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_levels_factor.R
\name{tranfofacto}
\alias{tranfofacto}
\title{La fonction "tranfofacto"}
\usage{
tranfofacto(var, codage)
}
\arguments{
\item{var}{de la forme data.frame$var}

\item{codage}{codage doit être une liste: list("new code"=c("old 1", "old 2"))}
}
\value{
un vecteur de facteurs res
}
\description{
...

tranfofacto: transforme les levels d'une variable factor
}
\author{
Elie
}
